Switch to single ocnice.config file
This merge changes the ocnice diagnostics to generate a single
configuration file, ocnice.config, that combines the configuration
from the default config file (python/MPAS-Analysis/config.analysis)
with the envornment variables set in run_*.csh.

These changes are intended to make is easier to debug cases where
configuation from ACME scripts are missing or set incorrectly, but
this goes unnoticed because default values from
python/MPAS-Analysis/config.analysis

Also, because a single file contains the full configuration, it is
easier to document the configuration that was used to run the
analysis scripts.