Fix __main__ directives in scripts/modules
For python files that can either be used as scripts or modules,
the 'if __name__ == "__main__":...' directives *should* be there
but they should ideally come after any function definitions and
there should be nothing other than imports that are outside of
either a function or the main directive.

This merge either removes the __main__ directive for some scripts
with no functions or moves it to the end for scripts that can also
function as modules.  Some content is included in the __main__
directive now that was outside of it before so that these scripts
would work properly as modules if they were to be used that way
in the future.